Performance: secondaryTables call percentage() (which is optimised)

+++ speeduino/corrections.cpp
@@ -367,7 +367,7 @@ uint16_t correctionAccel(void)
               //If CLT is less than taper min temp, apply full modifier on top of accelValue
               if ( currentStatus.coolant <= (int)(configPage2.aeColdTaperMin - CALIBRATION_TEMPERATURE_OFFSET) )
               {
-                uint16_t accelValue_uint = percentage(configPage2.aeColdPct, accelValue);
+                uint16_t accelValue_uint = percentage(configPage2.aeColdPct, (uint16_t)accelValue);
                 accelValue = (int16_t) accelValue_uint;
               }
               //If CLT is between taper min and max, taper the modifier value and apply it on top of accelValue
@@ -375,7 +375,7 @@ uint16_t correctionAccel(void)
               {
                 int16_t taperRange = (int16_t) configPage2.aeColdTaperMax - configPage2.aeColdTaperMin;
                 int16_t taperPercent = (int)((currentStatus.coolant + CALIBRATION_TEMPERATURE_OFFSET - configPage2.aeColdTaperMin) * 100) / taperRange;
-                int16_t coldPct = (int16_t) 100 + percentage( (100-taperPercent), (configPage2.aeColdPct-100) );
+                int16_t coldPct = (int16_t) 100 + percentage((uint8_t) (100U-taperPercent), (configPage2.aeColdPct-100U) );
                 uint16_t accelValue_uint = (uint16_t) accelValue * coldPct / 100; //Potential overflow (if AE is large) without using uint16_t (percentage() may overflow)
                 accelValue = (int16_t) accelValue_uint;
               }
@@ -432,7 +432,7 @@ uint16_t correctionAccel(void)
               //If CLT is less than taper min temp, apply full modifier on top of accelValue
               if ( currentStatus.coolant <= (int)(configPage2.aeColdTaperMin - CALIBRATION_TEMPERATURE_OFFSET) )
               {
-                uint16_t accelValue_uint = percentage(configPage2.aeColdPct, accelValue);
+                uint16_t accelValue_uint = percentage(configPage2.aeColdPct, (uint16_t)accelValue);
                 accelValue = (int16_t) accelValue_uint;
               }
               //If CLT is between taper min and max, taper the modifier value and apply it on top of accelValue
@@ -440,7 +440,7 @@ uint16_t correctionAccel(void)
               {
                 int16_t taperRange = (int16_t) configPage2.aeColdTaperMax - configPage2.aeColdTaperMin;
                 int16_t taperPercent = (int)((currentStatus.coolant + CALIBRATION_TEMPERATURE_OFFSET - configPage2.aeColdTaperMin) * 100) / taperRange;
-                int16_t coldPct = (int16_t)100 + percentage( (100 - taperPercent), (configPage2.aeColdPct-100) );
+                int16_t coldPct = (int16_t)100 + percentage((uint8_t) (100 - taperPercent), (uint16_t)(configPage2.aeColdPct-100) );
                 uint16_t accelValue_uint = (uint16_t) accelValue * coldPct / 100; //Potential overflow (if AE is large) without using uint16_t
                 accelValue = (int16_t) accelValue_uint;
               }
